requirements.txt
flask
# dotnet
python-dotenv==0.21.0
waitress==2.1.2
logging